requirements.txt
ipython==7.12.0
python-dotenv==0.11.0
requests==2.22.0
python-telegram-bot==12.4.2
PySocks==1.7.1